Move event handling to its own module
